# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AUD CBT Digital Therapeutic API
  description: |
    AUD(Alcohol Use Disorder) 디지털 치료 소프트웨어 v3.0 API 정본.

    입원 후 3개월 알코올 사용 장애 환자용 LLM 기반 인지행동치료(CBT) 앱의
    환자·의료진·내부 컴포넌트 간 통신 명세.

    범위: 대회 데모 + 사내 기술 검증. 실 환자 서비스 전 재구축 필요 항목은
    AUD_CBT_MVP_기능명세서 9장 Post-MVP 참조.
  version: 3.0.0
  contact:
    name: AUD CBT 개발팀
servers:
  - url: https://api.example.com/v1
    description: 운영 서버 (예시)
  - url: http://localhost:8000/v1
    description: 로컬 개발 서버
...
security:
  - BearerAuth: []
paths:
...
  /me/patient:
    get:
      tags: [Patient - Home]
      summary: P1 홈 화면 통합 응답
      description: |
        환자 본인의 홈 화면 정보를 한 번에 조회. 단주 일수, 현재 Week,
        다음 세션까지 남은 일수, 오늘 할 일, 다음 외래(next_outpatient_date),
        LLM 잠금 상태(llm_locked) 등.
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientHomeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # PATIENT APP - P2 CHECKIN
  # ============================================================
  /me/checkins:
    post:
      tags: [Patient - Checkin]
      summary: 오늘 체크인 제출 (하루 1회)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckinSubmit'
      responses:
        '201':
          description: 제출 성공. safety_classification에 위험 시 grade/event_type/next_action 포함.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckinResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: CHECKIN_ALREADY_SUBMITTED — 오늘 이미 체크인함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Patient - Checkin]
      summary: 체크인 목록 조회
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: 페이지네이션 응답
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Checkin' }
  /me/checkins/{checkin_id}:
    get:
      tags: [Patient - Checkin]
      summary: 체크인 단건 조회
      parameters:
        - $ref: '#/components/parameters/CheckinId'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Checkin' }
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Patient - Checkin]
      summary: 같은 날 체크인 수정 (24시간 이내)
      parameters:
        - $ref: '#/components/parameters/CheckinId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckinPatch'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckinResponse' }
        '409':
          description: 수정 시간 초과 (24h)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ============================================================
  # PATIENT APP - P3 CONVERSATION
  # ============================================================
...
  /provider/patients:
    post:
      tags: [Provider - D0]
      summary: D0 환자 신규 등록 + DischargeProfile 동시 저장
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DischargeProfileInput' }
      responses:
        '201':
          description: |
            등록 성공. registration_code(8자리)를 환자에게 SMS·종이 등으로 전달.
            primary_triggers는 내부 LLM이 normalized_triggers로 자동 변환.
          content:
            application/json:
              schema:
                type: object
                properties:
                  patient_id: { type: string, pattern: '^p_[a-z0-9]+$' }
                  registration_code:
                    type: string
                    pattern: '^[A-Z0-9]{8}$'
                  expires_at:
                    type: string
                    format: date-time
                    description: 등록 코드 만료 시각 (7일 후)
                  normalized_triggers:
                    type: array
                    items: { type: string }
                    example: [work_stress, social_pressure]
        '400':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Provider - D2]
      summary: 담당 환자 목록 조회 (v3.0은 정렬·필터 제한)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 페이지네이션 응답
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/PatientListItem' }
...
  /provider/patients/{patient_id}:
    get:
      tags: [Provider - D2]
      summary: 환자 상세 대시보드 (통합 응답)
      description: |
        외래 진료 시 5분 안에 환자 상태 파악용. v3.0에서는 모든 정보를
        이 응답 하나에 통합 (안전 이벤트·세션 이력·대화 로그·갈망 일지의 별도
        엔드포인트는 Post-MVP로 이동).
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: 환자 종합 정보
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientDetailDashboard' }
        '403':
          description: 담당 환자가 아님
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # PROVIDER PORTAL - D4 REASSESSMENT
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 단일 Access Token (v3.0 단순화).
        환자 24시간, 의료진 8시간 유효. Refresh Token·자동 갱신 없음.
        만료 시 재로그인 필요.
    InternalServiceAuth:
      type: apiKey
      in: header
      name: X-Internal-Service-Key
      description: |
        내부 API 전용. 각 백엔드 컴포넌트가 고유한 서비스 키 보유.
        90일 주기 키 교체. API 게이트웨이가 외부 클라이언트의 /internal/* 호출 차단.
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    PatientId:
      name: patient_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^p_[a-z0-9]+$'
        example: p_8f3a9b2c4d
...
    CheckinId:
      name: checkin_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: 401 인증 실패 (토큰 만료·서명 오류)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: 400 요청 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 404 리소스 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: 대문자 스네이크 케이스 에러 코드
              example: VALIDATION_ERROR
            message:
              type: string
              example: Request validation failed
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
            request_id:
              type: string
              format: uuid
    PaginatedEnvelope:
      type: object
      properties:
        items:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page: { type: integer }
            page_size: { type: integer }
            total_items: { type: integer }
            total_pages: { type: integer }
...
    SafetyGrade:
      type: string
      enum: [A, B]
      description: A=응급, B=비응급
    SafetyEventType:
      type: string
      enum: [suicide_risk, acute_intoxication, relapse, medication_stop, paws]
    NRSScore:
      type: integer
      minimum: 0
      maximum: 10

    # ---------- Patient Home (P1) ----------
    PatientHomeResponse:
      type: object
      properties:
        patient_id: { type: string }
        name: { type: string }
        sobriety_days: { type: integer }
        current_week: { type: integer, minimum: 1, maximum: 12 }
        days_to_next_session: { type: integer, nullable: true }
        today_tasks:
          type: object
          properties:
            checkin_pending: { type: boolean }
            session_today: { type: boolean }
        next_outpatient_date:
          type: string
          format: date
          nullable: true
        llm_locked:
          type: boolean
          description: 백엔드의 LLM 잠금 상태 (환자 앱이 차단 UI 노출 여부 판단에 사용)

    # ---------- Checkin (P2) ----------
    CheckinSubmit:
      type: object
      required: [mood_nrs, craving_nrs, sleep_hours, medication_records]
      properties:
        mood_nrs: { $ref: '#/components/schemas/NRSScore' }
        craving_nrs: { $ref: '#/components/schemas/NRSScore' }
        sleep_hours: { type: number, minimum: 0, maximum: 12 }
        medication_records:
          type: array
          items:
            type: object
            properties:
              medication_name: { type: string }
              taken: { type: boolean }
              side_effect_note: { type: string, nullable: true }
        free_note: { type: string, nullable: true }
    CheckinPatch:
      type: object
      description: CheckinSubmit의 부분 수정. 모든 필드 optional.
      properties:
        mood_nrs: { $ref: '#/components/schemas/NRSScore' }
        craving_nrs: { $ref: '#/components/schemas/NRSScore' }
        sleep_hours: { type: number, minimum: 0, maximum: 12 }
        medication_records:
          type: array
          items: { type: object }
        free_note: { type: string, nullable: true }
    Checkin:
      type: object
      properties:
        checkin_id: { type: string }
        date: { type: string, format: date }
        mood_nrs: { $ref: '#/components/schemas/NRSScore' }
        craving_nrs: { $ref: '#/components/schemas/NRSScore' }
        sleep_hours: { type: number }
        medication_records: { type: array, items: { type: object } }
        free_note: { type: string, nullable: true }
        submitted_at: { type: string, format: date-time }
    CheckinResponse:
      type: object
      properties:
        checkin: { $ref: '#/components/schemas/Checkin' }
        safety_classification:
          type: object
          nullable: true
          properties:
            grade: { $ref: '#/components/schemas/SafetyGrade' }
            event_type: { $ref: '#/components/schemas/SafetyEventType' }
            next_action:
              type: string
              enum: [llm_lock_p4, switch_resu, switch_soma, paws_resource, none]

    # ---------- Conversation (P3) ----------
...
    SafetyEvent:
      type: object
      properties:
        safety_event_id: { type: string, pattern: '^se_[a-z0-9]+$' }
        grade: { $ref: '#/components/schemas/SafetyGrade' }
        event_type: { $ref: '#/components/schemas/SafetyEventType' }
        detected_at: { type: string, format: date-time }
        source:
          type: string
          enum: [conversation_message, checkin_free_note, conversation_pattern]
        recommended_action:
          type: string
          enum: [llm_lock_p4, switch_resu, switch_soma, paws_resource, none]

    # ---------- Settings (P8) ----------
...
    SupportPersonInput:
      type: object
      required: [name, relationship, phone]
      properties:
        name: { type: string }
        relationship:
          type: string
          enum: [spouse, parent, sibling, child, friend, other]
        phone: { type: string }
...
    DischargeProfileInput:
      type: object
      required:
        - name
        - phone
        - date_of_birth
        - sex
        - discharge_date
        - diagnosis_severity
        - admission_days
        - medications
        - suicide_ideation_history
        - primary_triggers
        - sso
        - next_outpatient_date
      properties:
        name: { type: string }
        phone: { type: string }
        date_of_birth: { type: string, format: date }
        sex: { type: string, enum: [male, female, other] }
        discharge_date: { type: string, format: date }
        diagnosis_severity:
          type: string
          enum: [moderate, severe]
          description: DSM-5 중등도·중증
        admission_days: { type: integer, minimum: 1 }
        medications:
          type: array
          items: { $ref: '#/components/schemas/Medication' }
        comorbidities:
          type: array
          items:
            type: string
            enum: [depression, anxiety, insomnia, ptsd, bipolar, other]
        suicide_ideation_history:
          type: string
          enum: [none, past, during_admission, current]
        primary_triggers:
          type: object
          required: [raw_text]
          properties:
            raw_text:
              type: string
              description: 자유 텍스트 1~2줄. 백엔드 LLM이 정규화 태그로 변환.
        sso: { $ref: '#/components/schemas/SupportPersonInput' }
        next_outpatient_date: { type: string, format: date }
    Medication:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: naltrexone
          description: 'naltrexone | acamprosate | disulfiram | other'
        dose: { type: string, example: 50mg }
        frequency: { type: string, example: once_daily }
        note: { type: string, nullable: true }

    # ---------- D2 Patient list & detail ----------
    PatientListItem:
      type: object
      properties:
        patient_id: { type: string }
        name: { type: string }
        current_week: { type: integer }
        sobriety_days: { type: integer }
        last_active_at: { type: string, format: date-time, nullable: true }
        program_status:
          type: string
          enum: [active, completed, withdrawn]
        llm_locked: { type: boolean }
        unacknowledged_safety_events_count: { type: integer, minimum: 0 }
    PatientDetailDashboard:
      type: object
      properties:
        patient_id: { type: string }
        discharge_profile:
          type: object
          description: D0 정보 요약 (병명·약물·SSO·다음 외래일)
        progress:
          type: object
          properties:
            current_week: { type: integer }
            sobriety_days: { type: integer }
            medication_adherence_rate_30d: { type: number, format: float }
        recent_checkins_30d:
          type: array
          items: { $ref: '#/components/schemas/Checkin' }
        active_session:
          type: object
          nullable: true
        recent_sessions:
          type: array
          items:
            type: object
            properties:
              session_id: { type: string }
              week_number: { type: integer }
              ended_at: { type: string, format: date-time }
              summary: { $ref: '#/components/schemas/SessionSummary' }
        recent_safety_events:
          type: object
          description: v3.0에서 안전 이벤트 별도 조회·확인 처리 엔드포인트는 Post-MVP. 본 응답에 통합.
          properties:
            grade_a:
              type: array
              items: { $ref: '#/components/schemas/SafetyEvent' }
            grade_b:
              type: array
              items: { $ref: '#/components/schemas/SafetyEvent' }
        llm_lock_status:
          type: object
          properties:
            locked: { type: boolean }
            locked_at: { type: string, format: date-time, nullable: true }
            reason: { type: string, nullable: true }

    # ---------- Internal: Safety Classifier ----------
...
    SessionSummary:
      type: object
      description: 세션 요약기 출력. 다음 Week 세션 시작 시 시스템 프롬프트에 자동 주입.
      properties:
        session_completed_objectives:
          type: array
          items: { type: string }
        session_unaddressed_objectives:
          type: array
          items: { type: string }
        patient_key_insights:
          type: array
          items: { type: string }
        identified_triggers:
          type: array
          items:
            type: object
            properties:
              tag:
                type: string
                example: work_stress
              context: { type: string }
        assigned_homework: { type: string }
        emotional_tone:
          type: string
          enum: [engaged, resistant, low, volatile, neutral]
        next_session_handoff_notes: { type: string }
        safety_flags:
          type: array
          items:
            type: object
            properties:
              grade: { $ref: '#/components/schemas/SafetyGrade' }
              event_type: { $ref: '#/components/schemas/SafetyEventType' }
        generated_at: { type: string, format: date-time }
        model_used: { type: string, example: claude-sonnet-4-6 }
        generation_time_ms: { type: integer }

    # ---------- Internal: Output Filter ----------
...